% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msapi.R
\name{SP}
\alias{SP}
\title{SP}
\usage{
SP(
  query = "",
  Username = "",
  Method = "GET",
  acceptLanguage = "en",
  reenterPassword = FALSE,
  refreshTokens = FALSE,
  tokens_timeout_in_hours = "3",
  ...
)
}
\arguments{
\item{query}{The full URL for the query to send to the MS Office Graph or SharePoint API}

\item{Username}{Your email for authentication on the SharePoint server}

\item{Method}{Defaults to GET but also supports POST. MERGE support hasn't been written in yet.}

\item{acceptLanguage}{Defaults to en}

\item{reenterPassword}{Option to manually clear your password}

\item{refreshTokens}{Option to manually clear all tokens, for example, to connect from a different user account}

\item{tokens_timeout_in_hours}{Set number of hours before new tokens will be requested.
The default is 3, but regardless, new token requests will be handled as needed automatically.}
}
\value{
A data.frame with the complete, pagerized response to your query.
}
\description{
Simple, flexible REST API calls to MS SharePoint.
}
\details{
This function allows R to send normally constructed SharePoint queries
and return the response as a simple data.frame.

For more information about how to construct queries, see the SharePoint REST API documentation:
https://docs.microsoft.com/en-us/sharepoint/dev/sp-add-ins/complete-basic-operations-using-sharepoint-rest-endpoints

To further refine your queries, see the OData documentation:
https://www.odata.org/documentation/
}
\examples{
ListItems <- SP("https://{YOUR TENANT}.sharepoint.com/sites/_api/Web/Lists(guid'{THE ID OF YOUR LIST}')/Items", Username = "{YOUR AUTHENTICATION EMAIL}")


You can use pipes for a clearer view of your queries:

ListItems <- "https://{YOUR TENANT}.sharepoint.com/sites/_api/Web/Lists(guid'{THE ID OF YOUR LIST}')/Items" \%>\% SP(Username = "{YOUR AUTHENTICATION EMAIL}")
}
